modify tag_function
