+++ app/controllers/public/users_controller.rb
@@ -1,6 +1,7 @@
 class Public::UsersController < ApplicationController
   before_action :authenticate_user!
-  
+  before_action :is_matching_login_user, only: [:edit, :update]
+
   def index
     if params[:word]
       @users = User.search(params[:word])
@@ -12,7 +13,7 @@ class Public::UsersController < ApplicationController
   def show
     @user = User.find(params[:id])
     @users = User.all
-    @posts = current_user.posts
+    @posts = @user.posts
     # いいねした投稿を取得
     likes = Like.where(user_id: current_user.id).pluck(:post_id)
     @liked_posts = Post.find(likes)
@@ -21,7 +22,7 @@ class Public::UsersController < ApplicationController
   def edit
     @user = User.find(params[:id])
   end
-  
+
   def update
     @user = User.find(params[:id])
     if @user.update(user_params)
@@ -30,7 +31,7 @@ class Public::UsersController < ApplicationController
       render "edit"
     end
   end
-  
+
   def withdraw
     @user = User.find(current_user.id)
     # is_deletedカラムをtrueに変更することにより削除フラグを立てる
@@ -39,9 +40,16 @@ class Public::UsersController < ApplicationController
     flash[:notice] = "退会処理を実行いたしました"
     redirect_to new_user_registration_path
   end
-  
+
   private
   def user_params
     params.require(:user).permit(:name, :introduction, :profile_image)
+  end
+  
+    def is_matching_login_user
+    user = User.find(params[:id])
+    unless user.id == current_user.id
+      redirect_to posts_path
+    end
   end
 end
