--- app/javascript/menu.js
@@ -1,10 +1,0 @@
-$(function() {
-  $("ul.menu li").hover(
-    function() {
-      $(".menuSub:not(:animated)", this).slideDown();
-    },
-    function() {
-      $(".menuSub", this).slideUp();
-    }
-  );
-});
