--- app/javascript/page_top.js
@@ -1,8 +1,0 @@
-$(function() {
-$('#top a').on('click',function(event){
-  $('body, html').animate({
-    scrollTop:0
-  }, 800);
-  event.preventDefault();
-});
-});
