--- app/javascript/tab.js
@@ -1,9 +1,0 @@
-$(function() {
-  $('.tab').click(function(){
-    $('.tab-active').removeClass('tab-active');
-    $(this).addClass('tab-active');
-    $('.box-show').removeClass('box-show');
-    const index = $(this).index();
-    $('.tabbox').eq(index).addClass('box-show');
-  });
-});
