+++ app/models/tag.rb
@@ -1,7 +1,7 @@
 class Tag < ApplicationRecord
   has_many :post_tags,dependent: :destroy, foreign_key: 'tag_id'
   # タグは複数の投稿を持つ　それは、post_tagsを通じて参照できる
-  has_many :posts,through: :post_tags,dependent: :destroy
+  has_many :posts,through: :post_tags, dependent: :destroy
 
-  validates :name, uniqueness: true, presence: true
+  validates :name, presence: true, uniqueness: true
 end
