+++ app/views/admin/tags/index.html.erb
@@ -2,7 +2,16 @@
   <div class="row">
     <div class="col-7 mx-auto mt-5">
       <h1 class="text-center">タグ一覧</h1>
-      <table class="table table-borderless mt-3 ml-5">
+      <div class="mt-5">
+      <%= form_with model: @tag, url: admin_tags_path do |f| %>
+      <div class="row">
+        <%= f.label :name, "タグ追加", class: "col-3 col-form-label" %>
+          <%= f.text_field :name, class: "col-7 form-control col-6" %>
+          <%= f.submit "追加", class: "btn btn-success col-1 ml-1"%>
+        <% end %>
+      </div>
+    </div>
+      <table class="table table-borderless mt-5 ml-5">
         <% @tags.each do |tag| %>
           <tr>
             <td><%= tag.name %></td>
