+++ app/views/layouts/_sidebar.html.erb
@@ -57,12 +57,20 @@
         <i class="fas fa-tags"></i> タグ
         <% end %>
       </li>
-      <li>
-        <%= link_to "#", class: "text-decoration-none btn btn-outline-light btn-lg col-8" do %>
+      <li class="search-group">
+        <%= link_to "#", class: "text-decoration-none btn btn-outline-light btn-lg col-8 btn1" do %>
           <i class="fas fa-search"></i> 検索
         <% end %>
+        <!--検索フォーム-->
+       <div class="search-form col-10 mx-auto", style='display:none;'>
+         <%= form_with url: admin_users_path, local: true, method: :get, class: "input-group" do |form| %>
+            <%= form.text_field :word, class: "form-control justify-content-center" %>
+            <%= button_tag type: "submit", class: "btn btn-light ml-auto form-control col-3" do %>
+  	          <i class="fas fa-search"></i>
+            <% end %>
+          <% end %>
+       </div>
       </li>
-      <li><%= render "admin//users/search" %></li>
       <li>
         <%= link_to destroy_admin_session_path, method: :delete, class: "text-decoration-none btn btn-outline-light btn-lg col-8" do %>
         <i class="fas fa-sign-out-alt"></i> ログアウト
