+++ app/views/public/posts/new.html.erb
@@ -27,20 +27,28 @@
           <%= f.label :価格, class: "col-form-label col-3"%>
           <%= f.number_field :price, class: "col-9 form-control", placeholder: "半角で入力してください" %>
         </div>
-        <!--タグ機能-->
-        <div class="row mt-5">
-          <%= f.label :タグ, class: "col-form-label col-3" %>
-          <!--過去にユーザーが登録したタグ一覧-->
-          <%= f.collection_select :old_name, Tag.all, :id, :name, {include_blank: "タグを選択してください(複数選択はshift[command]を押しながら)"}, {class: "form-control col-9", multiple: true} %>
-          <p class="offset-md-3 mt-2">好きなタグがなければ下記のフォームより入力してください</p>
-          <!--新規でタグを登録するフォーム-->
-          <%= f.text_field :name, value: @tag_list, class: "offset-md-3 form-control", placeholder: ",(カンマ)で区切ると複数タグを登録できます"%>
-        </div>
         <!--表示非表示機能追加-->
         <!--地図情報入力フォーム-->
         <div class="row mt-5">
           <%= f.label :address, '場所', class: "col-form-label col-3" %>
           <%= f.text_field :address, class: 'col-9 form-control' %>
+        </div>
+         <!--タグ機能-->
+        <div class="row mt-5">
+          <%= f.label :タグ, class: "col-form-label col-3" %>
+          <!--過去にユーザーが登録したタグ一覧-->
+          <div class="d-flex flex-row flex-wrap p-2 border rounded col-9 mb-3">
+            <%= f.collection_check_boxes( :tag_ids, Tag.all, :id, :name) do |b| %>
+        	    <div class="mx-1">
+        	      <%= b.check_box %>
+        	      <%= b.label %>
+        	    </div>
+            <% end %>
+          </div>
+          <%#= f.collection_select :old_name, Tag.all, :id, :name, {include_blank: "タグを選択してください(複数選択はshift[command]を押しながら)"}, {class: "form-control col-9", multiple: true} %>
+          <p class="offset-md-3">タグがなければ下記のフォームより入力してください</p>
+          <!--新規でタグを登録するフォーム-->
+          <%= f.text_field :name, value: @tag_list, class: "offset-md-3 form-control", placeholder: ",(カンマ)で区切ると複数タグを登録できます"%>
         </div>
         <div class="row mt-5 d-flex justify-content-center">
           <%= f.submit "投稿する", class: "btn btn-outline-primary mr-5"%>
