+++ app/views/public/users/show.html.erb
@@ -29,17 +29,15 @@
   </div>
   <div class="row">
     <div class="col-10 mt-5", style="margin-left: 130px;">
-      <% if @user.id == current_user.id %>
-       <!--タブ部分-->
-        <ul class="tab-list list-unstyled d-flex ml-5">
-          <li class="tab tab-active text-center">
-            Posts
-          </li>
-          <li class="tab text-center">
-            Likes
-          </li>
-        </ul>
-      <% end %>
+     <!--タブ部分-->
+      <ul class="tab-list list-unstyled d-flex ml-5">
+        <li class="tab tab-active text-center">
+          Posts
+        </li>
+        <li class="tab text-center">
+          Likes
+        </li>
+      </ul>
     </div>
   </div>
   <div class="row">
